+++ packages/component-library/src/components/Section/index.ts
@@ -1,2 +1,2 @@
-export { default, } from './Section';
+export { default } from './Section';
 export type { SectionProps, SectionOverrides } from './Section';