+++ packages/component-library/src/components/Text/Text.tsx
@@ -153,7 +153,8 @@ function Text({ body, align, styles, variant, sidekickLookup, sx, renderNode, ..
         {...sidekick(sidekickLookup)}
         variant={variant}
         sx={{ textAlign: align, ...sx, ...styles?.root }}
-        {...props}>
+        {...props}
+      >
         {documentToReactComponents(body?.json, renderOptions({ links: body?.links, renderNode }))}
       </Root>
     </ErrorBoundary>