+++ packages/graphql-contentful-core/src/utils/createRichText.ts
@@ -1,4 +1,4 @@
- const createRichText = (text: string) => ({
+const createRichText = (text: string) => ({
   nodeType: 'document',
   data: {},
   content: [